requirements.txt
-illow==6.2.1
-ip==21.2.4
absl-py==0.15.0
affine==2.3.0
alabaster==0.7.12
antlr4-python3-runtime==4.9.3
anyio==2.0.2
argcomplete==1.12.3
argon2-cffi==21.1.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bokeh==2.4.2
brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click-plugins==1.1.1
click==8.0.3
cligj==0.7.2
cloudpickle==2.0.0
colorama==0.4.4
colorcet==3.0.0
commonmark==0.9.1
cycler==0.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash==2.0.0
dask==2022.1.0
datashader==0.13.0
datashape==0.5.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
distributed==2022.1.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
flake8==4.0.1
flask-compress==1.10.1
flask==2.0.2
flatbuffers==1.12
fsspec==2022.1.0
gast==0.3.3
gitdb==4.0.9
gitpython==3.1.27
glymur==0.9.6
google-auth-oauthlib==0.4.6
google-auth==2.3.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
heapdict==1.0.1
hydra-core==1.2.0
idna==3.3
imageio==2.9.0
imagesize==1.3.0
importlib-metadata==4.2.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython-genutils==0.2.0
ipython==7.30.1
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.17.2
jinja2==3.0.3
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.6.4
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-tabnine==0.0.24
jupyterlab-widgets==1.0.2
jupyterlab==3.2.4
kaleido==0.2.1
keras-preprocessing==1.1.2
keras==2.7.0
kiwisolver==1.3.2
llvmlite==0.38.0
locket==0.2.1
lxml==4.6.4
markdown==3.3.4
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.2.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.12.0
msgpack==1.0.3
multipledispatch==0.6.0
nbclassic==0.3.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numba==0.55.1
numpy==1.20.0
oauthlib==3.1.1
omegaconf==2.2.3
opencv-python==4.5.4.60
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
param==1.12.0
parso==0.7.1
partd==1.2.0
pathtools==0.1.2
pickleshare==0.7.5
pillow==9.0.0
pip==21.2.4
pipreqs==0.4.11
plotly==5.5.0
pluggy==0.13.1
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.9.0
py-cpuinfo==8.0.0
py==1.11.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pyct==0.4.6
pyflakes==2.4.0
pygments==2.10.0
pyparsing==3.0.7
pyrsistent==0.18.0
pytest-benchmark==3.4.1
pytest==4.6.11
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.3
pywavelets==1.1.1
pywin32==302
pywinpty==1.1.6
pyyaml==6.0
pyzmq==22.3.0
qtconsole==5.2.1
qtpy==1.11.2
random-username==1.0.2
rasterio==1.1.0
recommonmark==0.5.0
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.8
scikit-image==0.18.3
scipy==1.7.3
seaborn==0.11.2
send2trash==1.8.0
sentry-sdk==1.9.9
setproctitle==1.3.2
setuptools==58.0.4
shapely==1.7.1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
sphinx-autodoc-typehints==1.7.0
sphinx==2.4.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tblib==1.7.0
tenacity==8.0.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.7.0
tensorboardx==1.6
tensorflow-estimator==2.4.0
tensorflow==2.4.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.11.2
tomlkit==0.7.2
toolz==0.11.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchdatasets==0.2.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
ujson==5.1.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
xarray-spatial==0.3.1
xarray==0.20.2
xlsxwriter==3.0.2
yarg==0.1.9
zict==2.0.0
zipp==3.6.0